requirements.txt
pandas
mariadb
mysql
mysql-connector-python
python-dotenv